refactor: 컴포넌트 타입 types 파일로 변경
