+++ packages/stark-ui/src/components/Input/types.ts
@@ -1,0 +1,8 @@
+import type { IconProps } from 'starkds-icons';
+
+export type InputProps = {
+  error?: boolean;
+  helperText?: string;
+  rightIcon?: React.ReactElement<IconProps>;
+  label?: string;
+} & React.InputHTMLAttributes<HTMLInputElement>;
