+++ packages/stark-ui/src/components/Switch/types.ts
@@ -1,0 +1,4 @@
+export type SwitchProps = { size?: 'sm' | 'md' } & Omit<
+  React.InputHTMLAttributes<HTMLInputElement>,
+  'size'
+>;
